+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 ### Changed
 
 - Home page: reduce landscape banner image height to 40% on viewport widths less than 1100px.
+- Content grid: fetch collection cover image data only for collections included in the config.
 
 ### Fixed